+++ backend/app/controllers/SwimmerController.py
@@ -1,40 +1,23 @@
-from flask import request, Request, jsonify, abort #B added Request
-from app.models import db, Swimmer, Parent, Instructor, Lesson
-from app.run import app
+# app/controllers/SwimmerController.py
+from .. import db
+from ..models import Swimmer  # Assuming you have a model for Swimmer
 
-#Create, Read, Update, Delete Methods for Swimmer
-@app.route('/swimmers', methods=['POST'])
-def add_swimmer():
-    data = request.get_json()
-    #data = Request.get_json(force=False, silent=False, cache=True)
-    new_swimmer = Swimmer(name=data['name'], level=data['level'], special_needs=data.get('special_needs', None))
+def create_swimmer(data):
+    new_swimmer = Swimmer(**data)
     db.session.add(new_swimmer)
     db.session.commit()
-    return jsonify(new_swimmer.to_dict()), 201
+    return new_swimmer
 
-@app.route('/swimmers', methods=['GET'])
-def get_swimmers():
-    swimmers = Swimmer.query.all()
-    return jsonify([swimmer.to_dict() for swimmer in swimmers])
+def update_swimmer(swimmer_id, data):
+    swimmer = Swimmer.query.get(swimmer_id)
+    if swimmer:
+        for key, value in data.items():
+            setattr(swimmer, key, value)
+        db.session.commit()
+    return swimmer
 
-@app.route('/swimmers/<int:id>', methods=['GET'])
-def get_swimmer(id):
-    swimmer = Swimmer.query.get_or_404(id)
-    return jsonify(swimmer.to_dict())
-
-@app.route('/swimmers/<int:id>', methods=['PUT'])
-def update_swimmer(id):
-    swimmer = Swimmer.query.get_or_404(id)
-    data = request.get_json()
-    swimmer.name = data.get('name', swimmer.name)
-    swimmer.level = data.get('level', swimmer.level)
-    swimmer.special_needs = data.get('special_needs', swimmer.special_needs)
-    db.session.commit()
-    return jsonify(swimmer.to_dict())
-
-@app.route('/swimmers/<int:id>', methods=['DELETE'])
-def delete_swimmer(id):
-    swimmer = Swimmer.query.get_or_404(id)
-    db.session.delete(swimmer)
-    db.session.commit()
-    return jsonify({'message': 'Deleted'})
+def delete_swimmer(swimmer_id):
+    swimmer = Swimmer.query.get(swimmer_id)
+    if swimmer:
+        db.session.delete(swimmer)
+        db.session.commit()
